implemented move system with tests

+++ src/bitbishop/moves/move_execution.cpp
@@ -1,15 +1,15 @@
-#include <bitbishop/move_execution.hpp>
+#include <bitbishop/moves/move_execution.hpp>
 
 void MoveExecution::add(const MoveEffect& effect) { effects[count++] = effect; }
 
 void MoveExecution::apply(Board& board) const {
-  for (uint8_t i = 0; i < count; ++i) {
+  for (int i = 0; i < count; ++i) {
     effects[i].apply(board);
   }
 }
 
 void MoveExecution::revert(Board& board) const {
-  for (uint8_t i = count - 1; i >= 0; --i) {
+  for (int i = count - 1; i >= 0; --i) {
     effects[i].revert(board);
   }
 }
